requirements.txt
linkedin-scraper>=3.1.1
playwright
python-dotenv
pandas
openpyxl
openai
